pyproject.toml
[project]
name = "image-metadata-lab"
version = "0.1.0"
description = "Practice OpenCV image processing and EXIF (GPS/DateTime) metadata reading."
requires-python = ">=3.12"
dependencies = [
  "opencv-python>=4.13",
  "pillow>=12.1",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
  "jupyterlab>=4.4",
  "matplotlib>=3.10",
  "pytest>=9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]